make %e^X > a^Y and %e^X > a when a =/= %e.
Modify ordmexpt to make %e^X > a^Y and %e^X > a when a =/= %e.  This fixes some testsuite failures and creates a few more.  I'm not 100% certain that this change is compatible with all the requirements of an ordering function.

New file 'great` . This allows for looking for bugs in great